libcamera: framebuffer: Fix cookie arg type
The cookie has type `uint64_t`, but the constructor argument is of type
`unsigned int`. Fix that.

Unfortunately this is a public API and ABI change.

Fixes: d5b0cec83940 ("libcamera: framebuffer: Turn the cookie into a 64-bit integer")

+++ include/libcamera/framebuffer.h
@@ -58,7 +58,7 @@ public:
 		unsigned int length;
 	};
 
-	FrameBuffer(std::span<const Plane> planes, unsigned int cookie = 0);
+	FrameBuffer(std::span<const Plane> planes, uint64_t cookie = 0);
 	FrameBuffer(std::unique_ptr<Private> d);
 	virtual ~FrameBuffer() {}
